move endpoints for getting sets and tags by user to users controller
